Refactor specs
